edit

+++ 105_lab1.3.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 #include<math.h>
 int main()
-{   int a,b;
+{   unsigned long a,b;
 
     scanf("%d",&a);
     printf("%d",a%3);
